⚡️ Faster generate process for `mapToConstant`
The arbitrary `mapToConstant` used to rely on a linear scan to get the entry to be used to generate the element. It used to be sub-optimal for cases implying several entries (a normal scenario for this arbitrary).

Our new implementation relies on a dichotomic-lookup that will be way less costly to execute.

This change should have a visible impact when generating strings using the `string` arbitrary.