⚡️ Faster tokenizer of strings
Our `string` arbitrary starts its initialization by tokenizing known vulnerable strings into set of units (chars). The idea behind this tokenization process is to later generate vulnerable strings while generating entries with this arbitrary.

The process is the following:
- for each string known to be vulnerable, try to tokenize it with respect to the provided constraints on length and the unit arbitrary
- for each tokenizable string, add it to the bucket of potentially to be generated strings

This original tokenizer process was able to abide by constraints on length. Computed tokens were depending on the set of provided contraints on lengths and the arbitrary being considered. But this flexibility had a runtime cost we don't want to pay anymore. The tokenizer will stop trying to optimize on the lengths and will just tokenize for the requested arbitrary.