⚡️ Faster tokenizer of strings (#5387)
**Description**

<!-- Please provide a short description and potentially linked issues
justifying the need for this PR -->

Our `string` arbitrary starts its initialization by tokenizing known
vulnerable strings into set of units (chars). The idea behind this
tokenization process is to later generate vulnerable strings while
generating entries with this arbitrary.

The process is the following:
- for each string known to be vulnerable, try to tokenize it with
respect to the provided constraints on length and the unit arbitrary
- for each tokenizable string, add it to the bucket of potentially to be
generated strings

This original tokenizer process was able to abide by constraints on
length. Computed tokens were depending on the set of provided contraints
on lengths and the arbitrary being considered. But this flexibility had
a runtime cost we don't want to pay anymore. The tokenizer will stop
trying to optimize on the lengths and will just tokenize for the
requested arbitrary.

<!-- * Your PR is fixing a bug or regression? Check for existing issues
related to this bug and link them -->
<!-- * Your PR is adding a new feature? Make sure there is a related
issue or discussion attached to it -->

<!-- You can provide any additional context to help into understanding
what's this PR is attempting to solve: reproduction of a bug, code
snippets... -->

**Checklist** — _Don't delete this checklist and make sure you do the
following before opening the PR_

- [x] The name of my PR follows [gitmoji](https://gitmoji.dev/)
specification
- [x] My PR references one of several related issues (if any)
- [x] New features or breaking changes must come with an associated
Issue or Discussion
- [x] My PR does not add any new dependency without an associated Issue
or Discussion
- [x] My PR includes bumps details, please run `yarn bump` and flag the
impacts properly
- [x] My PR adds relevant tests and they would have failed without my PR
(when applicable)

<!-- More about contributing at
https://github.com/dubzzz/fast-check/blob/main/CONTRIBUTING.md -->

**Advanced**

<!-- How to fill the advanced section is detailed below! -->

- [x] Category: ⚡️ Improve performance
- [x] Impacts: Slight performance uplift, but way more to come thanks to
this change

<!-- [Category] Please use one of the categories below, it will help us
into better understanding the urgency of the PR -->
<!-- * ✨ Introduce new features -->
<!-- * 📝 Add or update documentation -->
<!-- * ✅ Add or update tests -->
<!-- * 🐛 Fix a bug -->
<!-- * 🏷️ Add or update types -->
<!-- * ⚡️ Improve performance -->
<!-- * _Other(s):_ ... -->

<!-- [Impacts] Please provide a comma separated list of the potential
impacts that might be introduced by this change -->
<!-- * Generated values: Can your change impact any of the existing
generators in terms of generated values, if so which ones? when? -->
<!-- * Shrink values: Can your change impact any of the existing
generators in terms of shrink values, if so which ones? when? -->
<!-- * Performance: Can it require some typings changes on user side?
Please give more details -->
<!-- * Typings: Is there a potential performance impact? In which cases?
-->