⚡️ Faster initialization of `string` with faster slices
When initializing the `string` arbitrary, we first check what are the known-to-be-vulnerable-strings that we could possibly generate for this instance.

As such we need to take any known-to-be-vulnerable-string and check if it could have been generated by the string builder we are planning to build. This operation is pretty costly as it implies a backtracking algorithm possibly failing and retrying multiple times.

The change introduced by this PR reduces the overhead linked to this computation by dropping part of the operations that were done twice.