✨ Add back strong unmapping capabilities to `string`
Reverting part of the capability-drop introduced by #5387.

When working on #5387, we dropped some unmapping capabilities on `string`. In the past `string` arbitrary used to be able to shrink even custom values and that even if it uses a pretty uncommon dictionary of units. With #5387 we restricted the unmapping capabilities a bit by dropping some of these very advanced supports.

Now that we know that we can add this support back while preserving our performance uplift we add part of them back.

In other words, before #5387 the following arbitrary could have shrunk `...___...`:

```ts
fc.string({ maxLength: 3, unit: fc.constantFrom('._', '_...', '_._.', '_..', '.', '.._.', '__.', '....', '..', '.___', '._..', '__', '_.', '___', '.__.', '__._', '._.', '...', '_', '.._', '..._', '.__', '_.._', '_.__', '__..') })
```

After the PR #5387, he was not able anymore.

But we are adding back such support! Which will be mostly invisible for all our users :joy: