👷 Preview builds using `pkg.pr.new`
The GitHub Application `pkg.pr.new` is responsible to help into previewing packages as they get developped. It could be seen as a possible alternative to the current codesandbox option that we have been using until now.

This PR aims into giving it a try to see how it behaves against our monorepo.