⚡️ Faster instantiation of internet-related arbitraries
These arbitraries used to be relatively costly to instantiate and re-instantiate. While generally not a huge concern, when used in multiple properties and arbitraries this small cost (around 10ms measured for some of them) could be considered as far too much.

Tis PR is responsible to drop this time for later re-instantiations of the same internet-related arbitrary (and also other internet-related aribitraries).