🐛 Make `stringify` a bit more resilient to poisoning
While we may still miss lots of poisoning cases, we try to be closer to a fully poisoning free implementation for `stringify`. There are still problems and risks of poisoning in the implementation but we dropped some of them with this PR.

In theory it should make tests easier to troubleshoot in case the test failed because of a poisoning it introduced itself.