📝 Do not display success count
We used to display the number of success for each day of our Advent of PBT. Unfortunately the service we have been trying on is missing lots of points as it tends to timeout on a regular basis. As such it makes it not really usable.

I prefer not showing this stat this year. Maybe I could find an alternative for my next advent.