📝 Add `typespec-fast-check` to ecosystem page (#5564)
**Description**

I've been working on the
[`typespec-fast-check`](https://github.com/TomerAberbach/typespec-fast-check)
package for a while and it's in a pretty usable state now, so I figured
I'd add it to the ecosystem documentation page. Check out the [OpenAI
example mentioned in the
README](https://github.com/TomerAberbach/typespec-fast-check#examples)!

**Checklist**

- [x] The name of my PR follows [gitmoji](https://gitmoji.dev/)
specification
- [x] My PR references one of several related issues (if any)
- [x] New features or breaking changes must come with an associated
Issue or Discussion
- [x] My PR does not add any new dependency without an associated Issue
or Discussion
- [x] My PR includes bumps details, please run `yarn bump` and flag the
impacts properly
- [x] My PR adds relevant tests and they would have failed without my PR
(when applicable)

**Advanced**

- [x] Category: 📝 Add or update documentation
- [x] Impacts: None

---

I'm assuming no bumps are needed because this doesn't affect any
published package.